requirements.txt
# Core web + API
flask
flask-cors
requests

# Optional for dashboard auto-refresh and sockets (if you extend)
flask-socketio

# For testing
pytest

# For Raspberry Pi hardware GPIO (only works on Pi, not Windows/Linux dev)
RPi.GPIO; sys_platform == 'linux'

# For serial comms with GSM/GPS (only needed when SIMULATE=false in gps_gsm_client.py)
pyserial